Dead Players and Enemies implemented
Animations are super scuffed though

Big issue is if someone dies while the round is over walking animation becomes super bad.
