+++ Assets/Scripts/Abstract Classes/EnemyClass.cs
@@ -43,8 +43,10 @@ public abstract class EnemyClass : EntityClass
 
     public override IEnumerator Die()
     {
-        yield return StartCoroutine(MoveToPosition(myTransform.position + new Vector3(10, 0, 0), 0, 0.8f));
+        int runDistance = 10;
+        yield return StartCoroutine(MoveToPosition(myTransform.position + new Vector3(runDistance, 0, 0), 0, 0.8f));
         CombatManager.Instance.RemoveEnemy(this);
+        isDead = true;
         this.gameObject.SetActive(false);
     }
