+++ Assets/Scripts/Abstract Classes/EntityClass.cs
@@ -52,14 +52,52 @@ public abstract class EntityClass : SelectClass
     }
 
 
-    public virtual void TakeDamage(int damage)
+    public virtual void TakeDamage(EntityClass source, int damage)
     {
         health = Mathf.Clamp(health - damage, 0, MAX_HEALTH);
         healthBar.setHealth(health);
+        float percentageDone = 1; //Testing different powered knockbacks
+        if (Health != 0)
+        {
+            percentageDone = Mathf.Clamp(damage / (float) Health, 0f, 1f);
+        }
+        if (isDead) return;
+        StartCoroutine(PlayHitAnimation(source, this, percentageDone));
+    }
+
+    private IEnumerator PlayHitAnimation(EntityClass origin, EntityClass target, float percentageDone)
+    {
+        if (isDead) yield break;
+
+        yield return StartCoroutine(StaggerEntities(origin, target, percentageDone));
         if (health <= 0)
         {
-           StartCoroutine(Die());
+            yield return StartCoroutine(Die());
         }
+    }
+
+    /* 
+    Purpose: Staggers an entity back 
+    origin: The origin of the damage/attack is coming from
+    target: The target being staggered back
+    percentageDone: Percentage health done to the target
+     */
+    private IEnumerator StaggerEntities(EntityClass origin, EntityClass target, float percentageDone)
+    {
+        Vector3 directionVector = target.myTransform.position - origin.myTransform.position;
+
+        Vector3 normalizedDirection = directionVector.normalized;
+        float staggerPower = StaggerPowerCalculation(percentageDone);
+        yield return StartCoroutine(target.StaggerBack(target.myTransform.position + normalizedDirection * staggerPower));
+    }
+
+    //Calculates the power of the stagger based on the percentage health done
+    private float StaggerPowerCalculation(float percentageDone)
+    {
+        float minimumPush = 0.8f;
+        float pushSlope = 1.8f;
+        float percentageUntilMaxPush = 1f / 3f; //Reaches Max push at 33% hp lost
+        return minimumPush + pushSlope * Mathf.Clamp(percentageDone / percentageUntilMaxPush, 0f, 1.5f);
     }
 
     /*
@@ -81,6 +119,7 @@ public abstract class EntityClass : SelectClass
         Vector3 diffInLocation = destination - originalPosition;
 
         if ((Vector2) diffInLocation == Vector2.zero) yield break;
+        if (isDead) yield break;
 
         float distance = Mathf.Sqrt(diffInLocation.x * diffInLocation.x + diffInLocation.y * diffInLocation.y);
         float maxProportionTravelled = (distance - radius) / distance;
