+++ Assets/Scripts/Entities/Jackie.cs
@@ -118,14 +118,4 @@ public class Jackie : PlayerClass
         return null;
     }
 
-
-
-    // Update is called once per frame
-    void Update()
-    {
-        if (Input.GetKeyDown(KeyCode.Space))
-        {
-            TakeDamage(5);
-        }        
-    }
 }
